Add dependency tree for Rust projects (#931)
* Add dependency tree for Rust projects

Fill the dependency tree for Rust projects if a Cargo.lock file is
available. The information to construct the tree is available in the
lock file, so the dependency tree construction is simply a transformation.

There does not seem to be a specification for Cargo.lock files, so the
implmentation makes assumptions of its structure based on available
examples. The assumptions should hold reasonably well because the files
should always be generated by Cargo.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

* Handle Windows line endings for Rust dependencies

Handle Windows-style line endings parsing Cargo.lock files to construct
dependency trees for Rust projects. The patterns for parsing the
sections from the Cargo.lock file will take either Linux- or
Windows-style line endings into account.

Add a test to verify that Cargo.lock files with Windows-style line
endings can be parsed.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

* Handle Cargo.lock files without footer

The pattern to find package sections in Cargo.lock files does not match
the last package if there is no metadata section at the end of the file.
Fix the pattern to account for a package section followed by EOF.

Multiple Cargo.lock files in a project may produce a dependency list
with duplicate entries, which is not compatible with the CycloneDX
specificaiton. Remove duplicate dependency entries.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

* Handle undefined packages in Rust dependency trees

In case a package is listed as a dependency in a Cargo.lock-file but is
not otherwise defined as a package, it is treated as an error. Continue
gracefully with an error message instead of failing when such a
dependency is encountered.

Add a test to make sure the function does not fail, with a description
of how it has been decided to work for now.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

* Make sure last package of Cargo.lock is included

In certain situations, the components list will miss the last package in
a Cargo.lock-file. A package is only added to the component list when
the Cargo.lock's next package is found. But when end-of-file is
encountered, the last package may be missed.

Fix the issue by adding the current pacakge to the list when the
end-of-file has been reached. Also add a very small test case to make
sure the implementation will include the last package.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

* Remove duplicate BOM-refs in Rust dependency list

If multiple different Cargo.lock-files lists a package of the same
version, there may occur duplicate BOM-references in the dependencies
list.

Merge the dependency lists of packages with identical BOM-refs.

The function `mergeDependencies` cannot be used to merge the lists,
because Rust has no "parent component" which the
`mergeDependencies`-function requires. Therefore it will be implemented
similarly, but not identically. The `mergeDependencies`-function can be
used when the `metadata.component` object of the BOM is created.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

* Add fallback metadata.component for Rust projects

A metadata.component object is needed to make use of the function
`mergeDependencies`. Create a reasonable metadata.component object from
the information available and remove the duplicate implementation of
`mergeDependencies`. Solving issue #939 is meant to find the information
from Cargo.toml-files instead.

Fix a test failing on Windows. The test always replaced `\n` with `\r\n`
causing the test on Windows to produce invalid newlines. The fix will
replace both Windows and Linux newlines with Windows newlines for the
test data.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

* Run gen-types after rebase

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

---------

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>