Add evidence for Cargo.lock parsed components (#960)
* Add evidence for Cargo.lock parsed components

Evidence for as to why the component was identified as part of the
project can be provided if the component was identified in the
Cargo.lock-file. Update the Cargo.lock component parser to also insert
the evidence, describing how it was found.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

* Lower Cargo.toml evidence confidence to 0.5

The confidence was previously set to 0.7 but was deemed to high due to
being a single source of information. Change the confidence to be lower
than that for Cargo.lock-files, which is considered being an additional
source of information.

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>

---------

Signed-off-by: Lucas Ljungberg <lucasljungberg@gmail.com>